Generalize the cats-vs-dogs script to a generic image classifier training script
